Apply Updates to Vapor Template
This is based on the latest changes within https://github.com/vapor/template/tree/main
