+++ .dockerignore
@@ -1,2 +1,3 @@
 .build/
 .swiftpm/
+.git/
