+++ Package.swift
@@ -16,7 +16,7 @@ let package = Package(
         .package(url: "https://github.com/vapor/apns.git", from: "1.0.0"),
     ],
     targets: [
-        .target(
+        .executableTarget(
             name: "App",
             dependencies: [
                 .product(name: "Vapor", package: "vapor"),
@@ -26,18 +26,6 @@ let package = Package(
                 .product(name: "S3", package: "aws-sdk-swift"),
                 .product(name: "LeafMarkdown", package: "leaf-markdown"),
                 .product(name: "APNS", package: "apns"),
-            ],
-            swiftSettings: [
-                // Enable better optimizations when building in Release configuration. Despite the use of
-                // the `.unsafeFlags` construct required by SwiftPM, this flag is recommended for Release
-                // builds. See <https://github.com/swift-server/guides/blob/main/docs/building.md#building-for-production> for details.
-                .unsafeFlags(["-cross-module-optimization"], .when(configuration: .release)),
-            ]
-        ),
-        .executableTarget(
-            name: "Run",
-            dependencies: [
-                .target(name: "App")
             ]
         ),
         .testTarget(
