+++ Procfile
@@ -1,2 +1,2 @@
-release: Run migrate -y
-web: Run serve --env production --hostname 0.0.0.0 --port $PORT
+release: App migrate -y
+web: App serve --env production --hostname 0.0.0.0 --port $PORT
